--- a/.changeset/empty-starfishes-leave.md
+++ b/.changeset/empty-starfishes-leave.md
@@ -1,0 +1,5 @@
+---
+"stylelint": patch
+---
+
+Fixed: `hue-degree-notation` performance