--- a/lib/rules/hue-degree-notation/index.js
+++ b/lib/rules/hue-degree-notation/index.js
@@ -24,6 +24,10 @@ const meta = {
 const HUE_FIRST_ARG_FUNCS = ['hsl', 'hsla', 'hwb'];
 const HUE_THIRD_ARG_FUNCS = ['lch'];
 const HUE_FUNCS = new Set([...HUE_FIRST_ARG_FUNCS, ...HUE_THIRD_ARG_FUNCS]);
+const HAS_HUE_COLOR_FUNC = new RegExp(
+	`\\b(?:${[...HUE_FIRST_ARG_FUNCS, ...HUE_THIRD_ARG_FUNCS].join('|')})\\(`,
+	'i',
+);
 
 /** @type {import('stylelint').Rule} */
 const rule = (primary, _secondaryOptions, context) => {
@@ -36,6 +40,8 @@ const rule = (primary, _secondaryOptions, context) => {
 		if (!validOptions) return;
 
 		root.walkDecls((decl) => {
+			if (!HAS_HUE_COLOR_FUNC.test(decl.value)) return;
+
 			let needsFix = false;
 			const parsedValue = valueParser(getDeclarationValue(decl));