--- a/lib/rules/hue-degree-notation/index.js
+++ b/lib/rules/hue-degree-notation/index.js
@@ -25,7 +25,7 @@ const HUE_FIRST_ARG_FUNCS = ['hsl', 'hsla', 'hwb'];
 const HUE_THIRD_ARG_FUNCS = ['lch'];
 const HUE_FUNCS = new Set([...HUE_FIRST_ARG_FUNCS, ...HUE_THIRD_ARG_FUNCS]);
 const HAS_HUE_COLOR_FUNC = new RegExp(
-	`\\b(?:${[...HUE_FIRST_ARG_FUNCS, ...HUE_THIRD_ARG_FUNCS].join('|')})\\(`,
+	`\\b(?:${[...HUE_FUNCS].join('|')})\\(`,
 	'i',
 );