<!-- Each pull request must be associated with an open issue unless it's a documentation fix. If a corresponding issue does not exist, please create one so we can discuss the change first. -->

<!-- Please answer the following. We close pull requests that don't. -->

> Which issue, if any, is this issue related to?

Ref https://github.com/stylelint/stylelint/issues/6869

> Is there anything in the PR that needs further explanation?

Mean: 120.86200307142857 ms
Deviation: 19.25088932105957 ms

To

Mean: 85.13097773333334 ms
Deviation: 8.810119272059223 ms